chore: Update import path for Cities data in AirQoMap.jsx and index.jsx
