+++ platform/src/common/components/Map/data/Cities.js
