+++ platform/src/common/components/Map/data/countries.json
